[front/components/skill_builder] - refactor: adjust SkillBuilder components organization
 - Moved `SkillBuilderVersionComparisonFooter` to a separate module for better separability and potential reuse
 - Removed unused imports and dependencies from `SkillBuilderVersionComparisonBanner.tsx` to clean up the codebase
 - Streamlined `SkillBuilder.tsx` imports to align with the new component structure
